colocate Issue logic with defined errors
