+++ src/issues.ts
@@ -1,4 +1,31 @@
-import {  Issues } from './util';
+type Params<T> = T extends (... args: infer T) => any ? T : never;
+type MessageVariable = string | number | boolean | null;
+
+class Issue extends Error {
+  constructor(message: string){
+    super(message);
+    // deletes first line of stack trace (helper call)
+    this.stack = this.stack!.replace(/\n.+/, "");
+  }
+
+  warn = () => { console.warn(this.message) }
+  throw = (): never => { throw this }
+}
+
+function Issues
+  <O extends BunchOf<(...args: MessageVariable[]) => string>>
+  (register: O){
+  
+  const Library = {} as any;
+
+  for(const name in register)
+    Library[name] = () => 
+      new Issue(register[name].apply(null, arguments as any));
+
+  return Library as {
+    readonly [P in keyof O]: (...args: Params<O[P]>) => Issue
+  };
+}
 
 export default Issues({
   ContextNotFound: (name) =>
